+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Use user's locale for date formatting
 - Improve layout/design of transaction import dialog
+- Split value column into credit/debit columns in transaction import dialog
 
 ## [0.6.0] - 2022-10-12